chore: Clean up old Activity panel code
