+++ src/app/Scenes/Activity/Activity.tsx
@@ -1,65 +1,67 @@
-import { ActionType, OwnerType } from "@artsy/cohesion"
-import { ClickedActivityPanelTab } from "@artsy/cohesion/dist/Schema/Events/ActivityPanel"
-import { Tabs } from "@artsy/palette-mobile"
-import { ActivityContainer } from "app/Scenes/Activity/ActivityContainer"
-import { NewActivityScreen } from "app/Scenes/Activity/NewActivityScreen"
-import { goBack } from "app/system/navigation/navigate"
+import { DEFAULT_HIT_SLOP, MoreIcon, Screen, Separator, Touchable } from "@artsy/palette-mobile"
+import { useActionSheet } from "@expo/react-native-action-sheet"
+import { ActivityContent, ActivityContentPlaceholder } from "app/Scenes/Activity/ActivityContent"
+import { ActivityScreenStore } from "app/Scenes/Activity/ActivityScreenStore"
+import { ActivityHeader } from "app/Scenes/Activity/components/ActivityHeader"
+import { useMarkAllNotificationsAsRead } from "app/Scenes/Activity/hooks/useMarkAllNotificationsAsRead"
+import { goBack, navigate } from "app/system/navigation/navigate"
+import { Suspense } from "react"
 
-import { useFeatureFlag } from "app/utils/hooks/useFeatureFlag"
-import { ProvideScreenTrackingWithCohesionSchema } from "app/utils/track"
-import { screen } from "app/utils/track/helpers"
-import { OnTabChangeCallback } from "react-native-collapsible-tab-view"
-import { useTracking } from "react-tracking"
-
-export const Activity = () => {
-  const enableNewActivityPanelManagement = useFeatureFlag("AREnableNewActivityPanelManagement")
-
-  const tracking = useTracking()
-
-  const handleTabPress: OnTabChangeCallback = (data) => {
-    tracking.trackEvent(tracks.clickedActivityPanelTab(data.tabName))
-  }
-
-  if (enableNewActivityPanelManagement) {
-    return (
-      <ProvideScreenTrackingWithCohesionSchema
-        info={screen({ context_screen_owner_type: OwnerType.activities })}
-      >
-        <NewActivityScreen />
-      </ProvideScreenTrackingWithCohesionSchema>
-    )
-  }
+export const Activity: React.FC = () => {
+  const type = ActivityScreenStore.useStoreState((state) => state.type)
+  const { showActionSheetWithOptions } = useActionSheet()
+  const { markAllNotificationsAsRead } = useMarkAllNotificationsAsRead()
 
   return (
-    <ProvideScreenTrackingWithCohesionSchema
-      info={screen({ context_screen_owner_type: OwnerType.activities })}
-    >
-      <Tabs.TabsWithHeader
+    <Screen>
+      <Screen.AnimatedHeader
+        onBack={goBack}
         title="Activity"
-        onTabChange={handleTabPress}
-        headerProps={{
-          onBack: goBack,
-        }}
-        headerHeight={40}
-      >
-        <Tabs.Tab name="All" label="All">
-          <Tabs.Lazy>
-            <ActivityContainer type="all" />
-          </Tabs.Lazy>
-        </Tabs.Tab>
-        <Tabs.Tab name="Alerts" label="Alerts">
-          <Tabs.Lazy>
-            <ActivityContainer type="alerts" />
-          </Tabs.Lazy>
-        </Tabs.Tab>
-      </Tabs.TabsWithHeader>
-    </ProvideScreenTrackingWithCohesionSchema>
-  )
-}
+        rightElements={
+          <Touchable
+            hitSlop={DEFAULT_HIT_SLOP}
+            onPress={() => {
+              showActionSheetWithOptions(
+                {
+                  options: ["Mark all as read", "Edit Alerts", "Edit Follows", "Cancel"],
+                  cancelButtonIndex: 3,
+                  useModal: true,
+                },
+                (buttonIndex) => {
+                  switch (buttonIndex) {
+                    case 0:
+                      markAllNotificationsAsRead()
+                      break
+                    case 1:
+                      navigate("settings/alerts", {
+                        showInTabName: "profile",
+                      })
+                      break
+                    case 2:
+                      navigate("favorites", {
+                        showInTabName: "profile",
+                      })
+                      break
+                  }
+                }
+              )
+            }}
+          >
+            <MoreIcon fill="black100" accessibilityLabel="Notifications menu" top="2px" />
+          </Touchable>
+        }
+      />
 
-const tracks = {
-  clickedActivityPanelTab: (tabName: string): ClickedActivityPanelTab => ({
-    action: ActionType.clickedActivityPanelTab,
-    tab_name: tabName,
-  }),
+      <Screen.StickySubHeader
+        title="Activity"
+        separatorComponent={<Separator borderColor="black5" />}
+      >
+        <ActivityHeader />
+      </Screen.StickySubHeader>
+
+      <Suspense fallback={<ActivityContentPlaceholder />}>
+        <ActivityContent type={type} />
+      </Suspense>
+    </Screen>
+  )
 }
