+++ src/app/Scenes/Activity/ActivityItem.tsx
@@ -28,8 +28,6 @@ const NEW_ARTWORK_IMAGE_SIZE = 60
 
 export const ActivityItem: React.FC<ActivityItemProps> = memo(
   (props) => {
-    const enableNavigateToASingleNotification = useFeatureFlag("AREnableSingleActivityPanelScreen")
-    const enableNewActivityPanelManagement = useFeatureFlag("AREnableNewActivityPanelManagement")
     const enableBlurhash = useFeatureFlag("ARShowBlurhashImagePlaceholder")
 
     const markAsRead = useMarkNotificationAsRead()
@@ -49,150 +47,80 @@ export const ActivityItem: React.FC<ActivityItemProps> = memo(
         markAsRead(item)
       }
 
-      navigateToActivityItem(item, enableNavigateToASingleNotification)
+      navigateToActivityItem(item)
     }
 
-    const showAsRow = enableNewActivityPanelManagement && isPartnerOffer
+    const showAsRow = isPartnerOffer
 
-    if (enableNewActivityPanelManagement) {
-      return (
-        <TouchableOpacity activeOpacity={0.65} onPress={handlePress}>
-          <Flex flexDirection="row" alignItems="center" px={2}>
-            <Flex flex={1} mr={2}>
-              <Flex flexDirection="column" py={2}>
-                <Flex flexDirection={showAsRow ? "row" : "column"}>
-                  <Flex flexDirection="row" alignItems="center">
-                    {item.previewImages.map((image) => {
-                      if (!image.url) return null
+    return (
+      <TouchableOpacity activeOpacity={0.65} onPress={handlePress}>
+        <Flex flexDirection="row" alignItems="center" px={2}>
+          <Flex flex={1} mr={2}>
+            <Flex flexDirection="column" py={2}>
+              <Flex flexDirection={showAsRow ? "row" : "column"}>
+                <Flex flexDirection="row" alignItems="center">
+                  {item.previewImages.map((image) => {
+                    if (!image.url) return null
 
-                      return (
-                        <Flex
-                          key={image.url}
-                          mr={1}
-                          mb={1}
-                          accessibilityLabel="Activity Artwork Image"
-                          height={NEW_ARTWORK_IMAGE_SIZE}
-                          width={NEW_ARTWORK_IMAGE_SIZE}
-                        >
-                          <Image
-                            src={image.url}
-                            width={NEW_ARTWORK_IMAGE_SIZE}
-                            height={NEW_ARTWORK_IMAGE_SIZE}
-                            showLoadingState={!props.isVisible}
-                            blurhash={enableBlurhash ? image.blurhash : undefined}
-                          />
-                        </Flex>
-                      )
-                    })}
-
-                    {!!shouldDisplayCounts && (
-                      <Text
-                        variant="xs"
-                        color="black60"
-                        accessibilityLabel="Remaining artworks count"
+                    return (
+                      <Flex
+                        key={image.url}
+                        mr={1}
+                        mb={1}
+                        accessibilityLabel="Activity Artwork Image"
+                        height={NEW_ARTWORK_IMAGE_SIZE}
+                        width={NEW_ARTWORK_IMAGE_SIZE}
                       >
-                        + {remainingArtworksCount}
+                        <Image
+                          src={image.url}
+                          width={NEW_ARTWORK_IMAGE_SIZE}
+                          height={NEW_ARTWORK_IMAGE_SIZE}
+                          showLoadingState={!props.isVisible}
+                          blurhash={enableBlurhash ? image.blurhash : undefined}
+                        />
+                      </Flex>
+                    )
+                  })}
+
+                  {!!shouldDisplayCounts && (
+                    <Text
+                      variant="xs"
+                      color="black60"
+                      accessibilityLabel="Remaining artworks count"
+                    >
+                      + {remainingArtworksCount}
+                    </Text>
+                  )}
+                </Flex>
+
+                <Flex>
+                  {!!isPartnerOffer && (
+                    <PartnerOfferBadge notificationType={item.notificationType} />
+                  )}
+
+                  <Text variant="sm-display" fontWeight="bold">
+                    {item.headline}
+                  </Text>
+
+                  {!!isEditorial && <Text variant="xs">{item.message}</Text>}
+
+                  <Flex flexDirection="row" mt="1px">
+                    <ActivityItemTypeLabel notificationType={item.notificationType} />
+                    {!isPartnerOffer && (
+                      <Text variant="xs" mr={0.5}>
+                        {item.publishedAt}
                       </Text>
                     )}
-                  </Flex>
-
-                  <Flex>
-                    {!!isPartnerOffer && (
-                      <PartnerOfferBadge notificationType={item.notificationType} />
+                    {shouldDisplayExpiresInTimer(item.notificationType, item.item) && (
+                      <Flex ml="1px">
+                        <ExpiresInTimer item={item.item} />
+                      </Flex>
                     )}
-
-                    <Text variant="sm-display" fontWeight="bold">
-                      {item.headline}
-                    </Text>
-
-                    {!!isEditorial && <Text variant="xs">{item.message}</Text>}
-
-                    <Flex flexDirection="row" mt="1px">
-                      <ActivityItemTypeLabel notificationType={item.notificationType} />
-                      {!isPartnerOffer && (
-                        <Text variant="xs" mr={0.5}>
-                          {item.publishedAt}
-                        </Text>
-                      )}
-                      {shouldDisplayExpiresInTimer(item.notificationType, item.item) && (
-                        <Flex ml="1px">
-                          <ExpiresInTimer item={item.item} />
-                        </Flex>
-                      )}
-                    </Flex>
                   </Flex>
                 </Flex>
               </Flex>
             </Flex>
-            {!!item.isUnread && (
-              <Flex
-                width={UNREAD_INDICATOR_SIZE}
-                height={UNREAD_INDICATOR_SIZE}
-                borderRadius={UNREAD_INDICATOR_SIZE / 2}
-                bg="blue100"
-                accessibilityLabel="Unread notification indicator"
-              />
-            )}
           </Flex>
-        </TouchableOpacity>
-      )
-    }
-    return (
-      <TouchableOpacity activeOpacity={0.65} onPress={handlePress}>
-        <Flex py={2} flexDirection="row" alignItems="center" px={2}>
-          <Flex flex={1}>
-            <Flex flexDirection="row">
-              <ActivityItemTypeLabel notificationType={item.notificationType} />
-
-              {!isPartnerOffer && (
-                <Text variant="xs" mr={0.5}>
-                  {item.publishedAt}
-                </Text>
-              )}
-            </Flex>
-
-            <Text variant="sm-display" fontWeight="bold">
-              {item.title}
-            </Text>
-
-            {!isPartnerOffer && <Text variant="sm-display">{item.message}</Text>}
-
-            {shouldDisplayExpiresInTimer(item.notificationType, item.item) && (
-              <ExpiresInTimer item={item.item} />
-            )}
-
-            <Spacer y={1} />
-
-            <Flex flexDirection="row" alignItems="center">
-              {item.previewImages.map((image) => {
-                if (!image.url) return null
-
-                return (
-                  <Flex
-                    key={image.url}
-                    mr={1}
-                    accessibilityLabel="Activity Artwork Image"
-                    height={NEW_ARTWORK_IMAGE_SIZE}
-                    width={NEW_ARTWORK_IMAGE_SIZE}
-                  >
-                    <Image
-                      src={image.url}
-                      width={ARTWORK_IMAGE_SIZE}
-                      height={ARTWORK_IMAGE_SIZE}
-                      showLoadingState={!props.isVisible}
-                    />
-                  </Flex>
-                )
-              })}
-
-              {!!shouldDisplayCounts && (
-                <Text variant="xs" color="black60" accessibilityLabel="Remaining artworks count">
-                  + {remainingArtworksCount}
-                </Text>
-              )}
-            </Flex>
-          </Flex>
-
           {!!item.isUnread && (
             <Flex
               width={UNREAD_INDICATOR_SIZE}
